code2/requirements.txt
flask
opencv-python
Pillow
numpy
mtcnn
keras-facenet
yt-dlp
tensorflow<2.11
